requirements.txt
streamlit
azure-cognitiveservices-speech
edge-tts
pydub
tqdm
pytest
